+++ src/network.cpp
@@ -467,6 +467,6 @@ void ServerThread::main_loop()
 		}
 
 		tick++;
-		next_logic = t0 + (tick + 1) * freq / TPS;
+		next_logic = t0 + ((Uint64)tick + 1) * freq / TPS;
 	}
 }
